Add respawn option
This commit let's you add an option to a package to have the entire process respawn after installing the package. This makes it possible to have `anms` update itself via a package that introduces new features used in other packages that is set to be updated/installed. 

Set `respawn: true` in the manifest file to enable this for a specific package.