Add forced refresh option
You can now use '-f' to force a recheck of the manifest file. This is useful if you f.eks fixed a condition on a computer that needs to be rechecked for possible installs. 

You can also add one or more '--refresh <mod>' to force-reinstall a package, even if it has not changed.